+++ pkg/providers/iam/provider.go
@@ -91,6 +91,7 @@ func (p *Subprovider) SDKDataSources() map[string]*schema.Resource {
 func (p *Subprovider) FrameworkResources() []func() resource.Resource {
 	return []func() resource.Resource{
 		NewIPAllowlistResource,
+		NewCIDRBlocksResource,
 	}
 }