+++ pkg/providers/iam/testdata/TestResCIDRBlocks/importable.tf
@@ -1,0 +1,13 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_cidr_blocks" "test" {
+  cidr_blocks = [
+    {
+      cidr_block = "128.5.6.6/24"
+      comments   = "test"
+      enabled    = true
+    }
+  ]
+}