+++ pkg/providers/iam/testdata/TestResCIDRBlocks/missing_cidr_block.tf
@@ -1,0 +1,12 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_cidr_blocks" "test" {
+  cidr_blocks = [
+    {
+      comments = "test"
+      enabled  = true
+    }
+  ]
+}