+++ pkg/providers/iam/testdata/TestResCIDRBlocks/update_single.tf
@@ -1,0 +1,12 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_iam_cidr_blocks" "test" {
+  cidr_blocks = [
+    {
+      cidr_block = "128.1.2.5/24"
+      enabled    = false
+    }
+  ]
+}